# Conflicts:
#	app/src/main/java/de/fraunhofer/fokus/OpenMobileNetworkToolkit/Iperf3/JSON/Sum.java
#	app/src/main/java/de/fraunhofer/fokus/OpenMobileNetworkToolkit/SettingPreferences/CarrierSettingFragment.java

+++ app/src/main/java/de/fraunhofer/fokus/OpenMobileNetworkToolkit/Iperf3/JSON/Timestamp.java
@@ -19,6 +19,6 @@ public class Timestamp {
     public String time;
     @SerializedName("timesecs")
     @Expose
-    public Integer timesecs;
+    public long timesecs;
 
 }
